@@ -5,15 +5,15 @@ homepage: SinaMN75.com
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.17.0"
+  sdk: '>=2.17.0'
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   url_launcher: ^6.0.15
-  get: ^4.6.1
+  get: ^4.6.3
   package_info_plus: ^1.4.2
   share_plus: ^4.0.3
   file_picker: ^4.5.1
@@ -23,7 +23,6 @@ dependencies:
   flutter_svg: ^1.0.3
   persian_tools: ^1.0.1
   group_button: ^5.0.1
-  smooth_page_indicator: ^1.0.0+2
   logger: ^1.1.0
   image_cropper: ^2.0.2
   cached_network_image: ^3.2.0
@@ -31,7 +30,7 @@ dependencies:
   just_audio: ^0.9.21
   syncfusion_flutter_pdfviewer: ^20.1.55-beta
   path_provider: ^2.0.10
-  flutter_downloader: ^1.7.3
+  flutter_downloader: ^1.7.4
   flutter_local_notifications: ^9.5.2
 #  firebase_messaging: ^11.2.12
 #  firebase_core: ^1.14.0
