@@ -1,56 +1,75 @@
 import 'package:utilities/utilities.dart';
 
-class UserRemoteDataSource {
+enum ProductDataSourceType {
+  products("Products"),
+  tutorials("Tutorials"),
+  projects("Projects"),
+  companies("Companies"),
+  tenders("Tenders"),
+  magazine("Magazine"),
+  ads("Ads"),
+  dailyPrice("DailyPrice"),
+  job("Job");
+
+  final String title;
+
+  const ProductDataSourceType(this.title);
+
+  @override
+  String toString() => name;
+}
+
+class ProductDataSource {
   final String baseUrl;
+  final ProductDataSourceType type;
 
-  UserRemoteDataSource({required this.baseUrl});
+  ProductDataSource({required this.baseUrl, required this.type});
 
-  Future<void> getMobileVerificationCodeForLogin({
+  Future<void> create({
     required final String url,
-    required final String mobile,
-    required final Function(GenericResponse<String>) onResponse,
+    required final ProductCreateUpdateDto dto,
+    required final Function(GenericResponse<ProductReadDto>) onResponse,
     required final Function(GenericResponse response) onError,
   }) async =>
       post(
-        url: "$baseUrl/user/GetMobileVerificationCodeForLogin",
-        body: GetMobileVerificationCodeForLoginDto(mobile: mobile, sendSms: true),
+        url: "$baseUrl/${type.title}",
+        body: dto,
         action: (Response response) => onResponse(GenericResponse.fromMap(response.body)),
         error: (Response response) => onError(GenericResponse.fromMap(response.body)),
       );
 
-  Future<void> verifyMobileNumberForLogin({
+  Future<void> update({
     required final String url,
-    required final String mobile,
-    required final String verificationCode,
-    required final Function(GenericResponse<UserReadDto>) onResponse,
+    required final ProductCreateUpdateDto dto,
+    required final Function(GenericResponse<ProductReadDto>) onResponse,
     required final Function(GenericResponse response) onError,
   }) async =>
-      post(
-        url: "$baseUrl/user/VerifyMobileForLogin",
-        body: VerifyMobileForLoginDto(mobile: mobile, verificationCode: verificationCode),
+      put(
+        url: "$baseUrl/${type.title}",
+        body: dto,
         action: (Response response) => onResponse(GenericResponse.fromMap(response.body)),
         error: (Response response) => onError(GenericResponse.fromMap(response.body)),
       );
 
-  Future<void> getProfile({
-    required final Function(GenericResponse<UserReadDto>) onResponse,
+  Future<void> read({
+    required final String url,
+    required final Function(GenericResponse<List<ProductReadDto>>) onResponse,
     required final Function(GenericResponse response) onError,
   }) async =>
       get(
-        url: "$baseUrl/user/GetProfile",
+        url: "$baseUrl/${type.title}",
         action: (Response response) => onResponse(GenericResponse.fromMap(response.body)),
         error: (Response response) => onError(GenericResponse.fromMap(response.body)),
       );
 
-  Future<void> updateProfile({
+  Future<void> readById({
     required final String url,
-    required final UpdateProfileParams params,
-    required final Function(GenericResponse<UserReadDto>) onResponse,
+    required final String id,
+    required final Function(GenericResponse<ProductReadDto>) onResponse,
     required final Function(GenericResponse response) onError,
   }) async =>
-      post(
-        url: "$baseUrl/user/UpdateProfile",
-        body: params,
+      get(
+        url: "$baseUrl/${type.title}/$id",
         action: (Response response) => onResponse(GenericResponse.fromMap(response.body)),
         error: (Response response) => onError(GenericResponse.fromMap(response.body)),
       );
